requirements.txt
openai>=1.0.0
requests>=2.0.0
pandas>=1.3.0
pyyaml>=6.0.0
appium-python-client>=2.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
allure-pytest>=2.9.0